Add sleep test, fix bug
slptest adds 3 processes to the event queue in ascending order of the
number of milliseconds to sleep, and then again in descending order.
The test passes if the messages printed to the console indicate they
awoke in the correct order.
This revealed a bug, which was fixed by initializing the `last' variable
to NULL.