New directory structure, recursive build
The directories under src/ have been moved into the root directory of
the project.  src/ is renamed kernel/.  bin/ and make/ are removed in
favour of building in the source directories, which is much simpler and
easier to maintain.

The build process is now recursive.  There is a makefile in each source
directory.  Generally these makefiles do little more than define the
objects to be built.

Additionally, some objcopy magic is performed in usr/ in order to link a
separate copy of klib.a into each user-space program.  First, the
program is compiled as usual to produce an object file.  This object
file is then (incrementally) linked with klib.a.  Then objcopy localizes
all symbols in the newly linked file except for the program's "main"
function, the name of which should correspond to the name of the source
file (so 'memtest.c' should define a function 'memtest' as its entry
point).  Since all symbols from klib.a have been stripped, the file
produced by objcopy should be safe to link with the kernel.