Build improvements
The linker script is now generated before linking.  This allows for the
names of input files to be discovered before the link and given to the
script as an INPUT (...) directive, instead of specifying the input
files statically in the linker command.

With the exception of usr, sub-makes no longer generate archive files.
Instead, they generate object files.  The main makefile then uses 'find'
to get a list of all the objects it needs to link.  An implication of
this is that sub-makes should only generate .o files if they're needed
for the link.