Improve CPU exception messages
Instead of guessing where the error code (if any) is on the stack, just
print out a portion of the stack and leave it to the reader to figure
out what's what.  This has the advantage of being less confusing
(usually) and printing some additional useful information--in
particular, the value of the instruction pointer when the exception
occurred.