Move paging_init() call, write protect ro sections
paging_init() is now called from mem_init(), and takes two parameters
for the start and end of the page heap.

Some linker variables are added to sections.ld and are used in
paging_init() to disable the R/W bits for pages containing read-only
data (i.e. the .text and .rodata sections).