Minor context switch optimization
The interrupt ID is written into the stack such that it will be in %eax
after the popa instruction.  This eliminates an unnecessary store/fetch,
as well as a move to %eax since the ID is the return value of
context_switch().  This can (and should) be done with 'rc' as well.

Note: gcc produces really bad code if optimization isn't turned on,
rendering this useless.