Simplify context switcher asm operands
The ISR_ENTRY macro is modified to accept a interrupt ID directly
instead of a string referring to an asm operand.  This required changing
the definitions for interrupt vectors to use preprocessor #defines
instead of an enum, so that they could be evaluated by the preprocessor
during the expansion of ISR_ENTRY.  This makes the various interrupt ID
operands unnecessary.