Use 'device_operations' struct
Drivers provide a 'device_operations' struct which lists the
device-specific routines to be called for open(), close() etc..  This
means that device drivers no longer need to expose the names of their
routines, and entries in the device table are less cumbersome to
initialize.

Also, instead of using a dummy routine which returns an error for
unsupported operations, members of the 'device_operations' struct may be
set to NULL.  The device-independent routines will check for this and
return an error if no routine is provided for a device.