Load kernel page directory after context switch
The kernel's page directory is loaded on every context switch into the
kernel.

Note: I had to hard-code the value of KERNEL_PAGE_OFFSET into the asm
constraint because of some stupidity with gcc (even though it agrees
that &_kernel_pgd and &KERNEL_PAGE_OFFSET are valid in an immediate
constraint, it complains that the difference of these values constitutes
an "impossible constraint").  I may have to replace KERNEL_PAGE_OFFSET
with a #define'd constant in the future.