Convert user pointers to physical addresses
Pointers from user address spaces are converted into physical addresses
for use in the kernel.

Also, the pmap_t type is defined for use with page directories.  This
name is shamelessly stolen from Mach, though pmap_t's in Mach contain
information in addition to the address of a page directory.