Separate functions for user/kernel process create
create_process() is split into functions create_kernel_process() and
create_user_process().  The helper function pcb_init() is defined in
order to capture some of the commonality between the two functions.

This will become more and more useful as differences between user and
kernel processes are implemented.  It was probably a bad idea to use a
general-purpose function to begin with.