New copy_*_userspace() functions
The copy_*_userspace() functions are rewritten to use temporary page
mappings instead of assuming the region is identity-mapped in the
kernel.

In order to accomplish this, a page table is set up in paging_init() for
use with temporary mappings, and the functions kmap_page() and
kunmap_page() are defined to perform the actual mappings.  kunmap_page()
doesn't actually map the requested address, but rather returns the new
virtual address at which the requested location can be accessed.  This
makes copy_through_userspace() much simpler, since we don't have to
worry about address conflicts.