Modify drivers to use temp mappings for user data
The keyboard and console drivers are modified to use copy_*_userspace()
and kmap_*() functions so that they no longer depend on user-space
memory being identity mapped in the kernel page tables.

In the case of the keyboard driver, the old buffering functionality is
removed.  This wasn't very useful, and greatly complicated the code.
The way the driver works now is to read into a buffer *only if there is
a read in progress* and then copy the buffer into user-space when it
fills up or the read completes.  This way the driver doesn't need to
hold on to a temporary mapping for the duration of a read.