RTC, system clock, time()
The RTC is read at boot to set the system clock (seconds since the
Epoch).  It is incremented every __TICKS_PER_SEC ticks.  This causes
a ridiculous amount of drift in QEMU.  I'm unsure if this is bad
emulation or an error on my part.

The POSIX time() and gmtime() functions are defined, including the
SYS_TIME syscall to get the value of the system clock.