Integrate kerboard driver into console driver
The keyboard driver was already interacting with the console driver;
this makes the connection explicit.

Ideally, a console would be represented as a virtual device on top of
separate keyboard and (text-mode) VGA drivers.  Current control-flow
limitations in Telos make this difficult, however.  So the console
driver drives both the physical keyboard and VGA interface.

The keyboard part of the driver buffers up to 512 characters.
Characters are delivered only after the buffer is full, a newline is
read, or EOF is reached.  The backspace character is understood by the
driver, so the backspace handling code is removed from tsh.