VFS: first pass & ramfs
Most of the VFS is ported at this point, but it's not well tested.
Mounting filesystems from block devices is not yet supported, since
block devices in general are not supported yet in Telos.

The reference/testing filesystem ("ramfs") is similar to tmpfs in that
it is not backed by a block device and claims memory as needed.  It uses
VFS inode structures as its own inodes, so a special flag is added to
the linux repertoire (MS_MEMFS) which prevents such inodes from being freed.