First pass at fork syscall
Implemented by copying the entire address space of the parent process,
since copy-on-write is not implemented (yet).  This currently works well
enough to print a message and then exit in the child.  Process
text/data/etc. segments are not copied, however, so this will break in a
number of ways.  In particular, only one of the parent/child can use
malloc after a fork, since the free list is shared.