execve, fcreate, binfs
execve is implemented, making it possible to do fork/exec for process
creation.

fcreate is a system call similar to syscreate, except that it runs an
executable from the filesystem rather than a function.

binfs is an in-memory filesystem which represents functions as
executable files.

With these in place, tsh no longer has to maintain its own
pseudo-filesystem.  It now fork/execs on files in binfs.