Various signal fixes/improvements/refactoring
Signal-related members are moved out of struct pcb and into their own
structure (sig_struct).

The SYS_SIGNAL syscall is removed.  libc can implement signal() as a
wrapper for sigaction(), for which there is already a syscall.

Previously ignoring vs. blocking signals was not handled properly.  This
is now fixed.  There is a sigset for pending signals and a sigset for
accepted signals (the "signal mask" in POSIX, but inverted).  A signal
is ignored if its handler is set to SIG_IGN.  Blocked signals become
pending when they are generated, but are not delivered until they are
unblocked.

Code for installing default signal handlers is moved into signal.c, and
the SIG_DFL behaviour is respected in sigaction.  There is a sig_exec()
function which resets the signal actions according to what POSIX
requires for the exec family of functions.