Major context switching changes
System calls always return to the calling process.  When a process wants
to block, it calls schedule(), which saves the kernel context and loads
the context of another process.  (Previously, blocking was accomplished
by returning from the syscall to another process.)

schedule() is now implemented as an assembly function (which calls
_schedule()) rather than using a software interrupt.

The message passing interface is removed because I didn't care to update
the code.  I plan to replace this with POSIX/SysV message passing
interfaces anyways, since these are better designed and more useful.