Macros to control type definitions
The <sys/type_macros.h> header defines macros of the form _<TYPE>_TYPE
for all types needed by libc.

e.g.

    #define _SIZE_T_TYPE unsigned long

Whenever a type needs to be defined in a header, it should first check
if _<TYPE>_DEFINED is defined, and only define the type if it is not.
Whenever a header defines a type, it should also define a macro of the
form _<TYPE>_DEFINED.

e.g.

    #include <sys/type_macros.h>

    #ifndef _SIZE_T_DEFINED
    #define _SIZE_T_DEFINED
    typedef _SIZE_T_TYPE size_t;
    #endif