waitid/waitpid/wait, orphans, reaping
waitid is a system call; waitpid and wait are implemented as wrappers.
waitid reaps zombies.

I'm not sure whether the wait functions are supposed to return only the
most recent status for a given process.  In this implementation, status
events are queued.

exit now orphans child processes to init, and the init process is
updated to repeatedly call wait to reap zombies.