Minimal tty driver layer; misc changes
The tty layer is based somewhat on the Linux tty layer.  Mostly the
function names and control flow are similar; the code itself is quite
different.  So far, the tty layer serves to separate buffering of input
data from the business of reading from a device.  Nothing fancy yet.

Also a bunch of other changes:
  * init is now a user-mode process
  * syscreate/fcreate are removed
  * "devfs" kluge removed
  * some code is moved around