Always pass string lengths as syscall arguments
When a string is passed as an argument to a system call, there should
always be an additional argument giving the length of the string.

Rationale: if the kernel knows the supposed length of a string, we can
easily check that it's valid in O(1) by calling vm_verify() and seeing if
there's a null byte where the user claims.  Otherwise we have to walk
the string, checking at each page boundary that the memory is mapped.

In the case of mount and execve, this pushes the number of arguments too
high, so a structure pointer is passed instead.  This is handled
transparently in libc.

Additionally, arbitrary limits on the number/length of arguments passed
to execve are now removed.  Environment variables are now copied as
well, though there is still no environ variable or *env() functions to
manipulate the environment.