Pass offset pointer to read/write file ops
The fs-specific read and write functions are passed a pointer to the file
position, and should update this value rather than file->f_pos.  This
allows the read/write functions to be used for pread, mmap, etc.  The
VFS read/write functions simply pass a pointer to file->f_pos.