vma_operations
Add a layer of indirection between the page fault handler and the
handling of a page fault.  This is meant to support memory-mapped files
and copy-on-write VMAs.  For mmap'd files, the op->map function would be
set to read a page of data from the file, and the op->writeback function
would be set to write a page back to the file.  For COW, the
op->write_perm function would be set to copy the page and set the
write-enable bit.

op->writeback currently isn't called anywhere.  It should be called from
del_pgdir() whenever a dirty page is encountered.  The function takes a
pointer and a length as arguments, so for efficiency del_pgdir() can
write multiple pages in a single call.