Move kernel stacks into kernel address space
Kernel stacks are mapped at the top of the kernel's address space, just
below the current page directory and temporary page table.

Old page directories are now cached so that they can be reused.  These
"husk" page directories map only the kernel (including a stack), so it
is not necessary to remap the kernel and allocate a new stack, page
tables, etc..

kmap_tmp_page() now panics if it cannot create a mapping, and is assumed
to succeed throughout the kernel.  A panic might be a bit extreme, but
if the kernel is trying to have > 1000 *temporary mappings*
simultaneously, it is a serious bug.