entry.S: save old current->esp on the stack
Save the old value of current->esp on the stack as part of save_all, and
restore it in restore_all.

This fixes a bug where a fault in-kernel would cause a general
protection fault on return_to_user, since the first fault clobbered
current->esp.  Now contexts effectively form a linked list, so it should
be possible to be interrupted arbitrarily many times (until the kernel
runs out of stack).