bio_vec: Scatter/gather block I/O
The bio_vec structure represents a list of blocks on a particular block
device.  A bio_vec may be read from/written to at arbitrary offests,
making it convenient for file I/O (the user doesn't have to worry about
crossing block boundaries).

An i_bio member is added to the inode structure.  If i_op->read is not
provided but i_bio is, then bio_read is used to complete reads.  For
many filesystems, this obviates the need to write a read function: it is
sufficient to store a file's block list in i_bio.

Also, various fixes for problems discovered while implementing ext2.