set_blocksize: configurable (buffer) block size
Block drivers no longer access buffers directly.  A block request is
given by { first_sector, nr_sectors, addr }, where first_sector is the
first sector of the request, nr_sectors is the number of sectors, and
addr is the memory address to read/write.  The sector size is fixed at
512 bytes, so devices with a different native sector size will have to
do some extra work.

Since the buffer block size no longer depends on the device sector size,
it is now possible to change the block size on a device.  The ext2 code
uses this to support different filesystem block sizes without adding
complexity to the code.