Use ext2 initrd
The first boot module is now the root device, and the kernel expects it
to have an ext2 (rev 0) filesystem on it.

The make target 'initrd.img' generates a suitable initrd using the
script 'scripts/mkfs.sh'.  It needs to run as root, since it mounts
filesystems and creates device nodes.

The init process currently mounts pseudo-filesystems on various
directories on the initrd that it expects to exist, so there are
some constraints on the initrd contents.  In the future, init can be
loaded from /init on the initrd, so that the kernel doesn't contain any
such assumptions (except to assume that /init exists).