Simplify build system
Use the MAKEFILES variable to automatically include $(topdir)/rules.mk
in all submakes.

Automatically call '$(MAKE) -C <submake> clean' for all $(submakes) in
the clean target (and analogously for distclean).  Automatically clean
all $(objects), so that $(clean) doesn't have to be defined in every
makefile.

For a submake with submakes of its own, this simple formula is all that
is required:

    all: [...] $(submakes)

    $(submakes):
            +$(call cmd,smake)
    .PHONY: $(submakes)