Move userspace onto initrd and purge binfs hack
binfs is finally gone.  All executables are now stored on a filesystem
and loaded by the kernel, rather than being compiled-in.

There's a bit of a hack in create_init.  Since exec assumes it's loading
the executable into an existing process, create_init is unable to reuse
this code to exec /bin/init directly.  Instead it creates a process from
a compiled-in function (as before), and this function calls exec on
/bin/init.