Refactor include/sys into include/telos
include/telos contains headers suitable for inclusion both in the kernel
and in libc.  include/sys is reserved for libc-specific headers.

The idea is that it should be possible to cp include/telos into the libc
source without modifications.  To this end, the headers in include/telos
are licensed under the 3-clause BSD license.