<telos/termios.h>
Define the termios structure and symbolic constants for <termios.h>.
Currently Telos doesn't do much with these.  Some infrastructure is
added to support per-terminal options, but there is currently no way to
change the termios parameters.

A generic tty ioctl is added to support tcgetattr.  The values returned
by this ioctl aren't very useful yet, but it can be used to implement
isatty since it will return an error on non-tty file descriptors.