Use i386-telos toolchain
We now use an OS-specific toolchain to build the kernel and initrd.  A
handy build repo for the toolchain/libc is forthcoming.

The kernel can still be built with the old i386-elf toolchain with a
simple edit to the makefile, but building the initrd without the telos
toolchain is significantly more difficult (various options must be
passed to the compiler/linker to get everything right).