pipes
Add a SYS_PIPE system call, taking 3 arguments: a location to store the
read-end of the pipe, a location to store the write-end of the pipe, and
a flags argument.  Currently the flags argument is unused.

In order to support the pipe implementation, the "flexbuf" data
structure is added.  This is a flexible buffer (meaning that memory is
allocated on demand) which may also be used as a queue.  The API is
similar to the file API.